// EntLib/src/app/Models/show.model.ts
export interface Show {
  id: string;
  title: string;
  showType: string;
  language: string;
  session: string;
  category: string;
  period: string;
  location: string;
  director: string;
  producer: string;
  cast: string;
  description: string;
  rating: string;
  posterUrl: string;
  trailerUrl: string;
  imdb: string;
}